requirements.txt
orjson
cn2an
coloredlogs
flatbuffers
numpy
packaging
protobuf
sympy
whl/Polygon3-3.0.9-cp311-cp311-win_amd64.whl
cnocr>=2.2.2.3
docopt
httpx
httpx_socks
loguru
opencv_python
Pillow
PyAutoGUI==0.9.53
pywinctl
pynput
pyuac
pywin32
questionary
tqdm
ultralytics
xpinyin
httpx_socks
flet
cryptography
pluggy
httpcore
pydantic
jsonschema
prompt-toolkit<=3.0.36,>=2.0
